Added additional tests'

+++ board.h
@@ -14,8 +14,8 @@ class Board { // TODO: make this entire class static
   public:
     // cell state methods
     Board(){};
-    Cell getCellState(unsigned short row, unsigned short col);
-    void setCellState(unsigned short row, unsigned short col, Cell state);
+    Cell getCellState(int row, int col);
+    void setCellState(int row, int col, Cell state);
     
     // board methods
     unsigned const short getHeight();
